% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pacmos_data.R, R/test-local-load.R
\name{load_pacmos_data}
\alias{load_pacmos_data}
\title{Load PACMOSData resources by dataset}
\usage{
load_pacmos_data(
  dataset = c("MESOMICS", "lungNENomics", "test"),
  local_dir = "eh-data"
)

load_pacmos_data(
  dataset = c("MESOMICS", "lungNENomics", "test"),
  local_dir = "eh-data"
)
}
\arguments{
\item{dataset}{One of \code{"MESOMICS"}, \code{"lungNENomics"}, or \code{"test"}.}

\item{local_dir}{Local folder containing prepared files. Default is \code{"eh-data"}.}
}
\value{
A named list of PACMOSData resources.

A named list of loaded PACMOSData resources.
}
\description{
Load PACMOSData resources by dataset

Load PACMOSData resources by dataset
}
\examples{
library(PACMOSData)

mesomics <- load_pacmos_data("MESOMICS")
lungnen <- load_pacmos_data("lungNENomics")
test <- load_pacmos_data("test")


}
