% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pacmos_data.R
\name{load_pacmos_data}
\alias{load_pacmos_data}
\title{Load PACMOSData resources by dataset}
\usage{
load_pacmos_data(dataset = c("MESOMICS", "lungNENomics", "test"))
}
\arguments{
\item{dataset}{One of \code{"MESOMICS"}, \code{"lungNENomics"}, or \code{"test"}.}
}
\value{
A named list of PACMOSData resources.
}
\description{
Load PACMOSData resources by dataset
}
